+++ util/fsmalloc.c
@@ -87,7 +87,7 @@ int fsmalloc_alloc(fsmalloc_t *fsmalloc, uintptr_t *addr, uint64_t count)
 }
 
 void fsmalloc_init(fsmalloc_t *fsmalloc, uintptr_t base_addr, uint64_t cell_size, uint64_t num_cells,
-                   bitarray_t *bitarr, uint64_t *words, uint64_t num_words)
+                   bitarray_t *bitarr, uint64_t *words, size_t num_words)
 {
     assert(num_words >= BITS_2_WORDS64(num_cells));
